Simplify the third regexp of File::Spec::Unix::canonpath()
At the second regexp, any path looking like './././xx' gets turned into
'./xx', so there is no need to look for several './' at the beginning of
the path anymore. This means that '(?:\./)+' can be simplified into '\./'.

This also makes it more clear why '././' is not turned into the empty
string by the third regexp even though it is not equal to './'.

Also remove the /s flag, as there is no '.' (sic) in the regexp.