Optimize File::Spec::Epoc::canonpath()
Since File::Spec::Epoc::canonpath() looks a lot like
File::Spec::Unix::canonpath(), it is easy to apply to the former the
optimizations devised for the latter.

The tests differing between File::Spec::Unix::canonpath() and
File::Spec::Epoc::canonpath() are also enabled by this commit, so that
it is easy to catch a behaviour change in the future.