regcomp.c: Add node POSIXA to optimizer
Commit 0658cddeb9feb16c427ac50f4000b008516b9958 added POSIXA and POSIXN,
but did not add handling of them to the optimizer.  The optimizer
defaults to assuming they can match every possible character under these
circumstances, so no bugs were introduced, but optimizations didn't
happen.

This commit copies the code that other similar nodes use, changing only
a very few things:
1) the loops that don't match any upper Latin1 characters are changed to
   only go through the possible ones it could match. i.e. 0 - 127, and
   to use the macro that only matches in that range.
2) It also adds UNI_TO_NATIVE() calls to compensate for EBCDIC.
3) The macro called in the loop is constructed based on the FLAGS
   parameter of the node, which has been set up to indicate which to use.

Even though only POSIXA and NPOSIXA are currently used, this commit is
set up to handle the complete set, taking advantage of the data
structure so that the regnode's number is offset from the base by the
enum of the character set.