regcomp.c: Merge quadruplicate code
There were 4 code sections (although one of them was just added) that
did essentially the same thing but for different regnodes.  A recent
commit enabled us to map a regnode to the macro that tests if something
belongs in it or not.  This commit takes advantage of that, and changes
all but the recently added section to do setup and then jump to that
section. which already used the macro mapping.